requirements.txt
flask==3.1.0
google-generativeai==0.8.4
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.1.0
